@@ -1,313 +1,349 @@
-//const utils = require('../../modules/utils.js');
+import { getDefaultSettings, initializeSettings, searchEngineList, addElements, createElement } from 'modules/utils.js';
 
-const defaultSettings = { modifiers: ["none"], default: "tdk", searchSelectedTextOnPopupShow: false };
-
-let settings = defaultSettings;
-
-async function loadSettings() {
-    settings = await browser.storage.local.get();
-
-    checkSettings();
-}
-
-loadSettings();
-
-let waitingBrowserActionToSendAWord = false;
-let objectToSendToBrowserAction = null;
-
-//browserAction portu
-let portbA = null;
+let settings = getDefaultSettings();
+let isWaitingForThePopupToLoadToSendAWord = false;
+let objectToBeSentToPopupWhenLoaded = null;
+let popupPort = null;
 
 //Sends meaning of the word that came from context menu.
-async function sendWordToTab(info, tab) {
-    let searchFrom = settings["default"];
-
-    if (searchFrom === "tdk") {
-        SearchFromTDK({ message: info.selectionText, isContextMenu: true }, { tab: tab });
-    } else if (searchFrom === "google") {
-        let text = await SearchFromGoogle(info.selectionText);
-
-        browser.tabs.sendMessage(tab.id, { type: "response_from_google", doc: text, isContextMenu: true });
-
-        text = null;
-    }
-
-    searchFrom = null;
+function sendContextMenuSelectedTextToTab(info, tab) {
+    browser.tabs.sendMessage(tab.id, { type: "search_from_context_menu", word: info.selectionText.trim().toLocaleLowerCase(), isContextMenu: true });
 }
 
-//async olursa browserAction açılmıyor.
-function ContextMenuClicked(info, tab) {
-    //Content script'e kutucuğu kapatmasını söylüyor.
-    browser.tabs.sendMessage(tab.id, { type: "SearchFromContextMenu" });
-
+//async olursa browserAction acilmiyor.
+function contextMenuClicked(info, tab) {
     if (info.menuItemId == "cm_onpopup") {
-        waitingBrowserActionToSendAWord = true;
+        isWaitingForThePopupToLoadToSendAWord = true;
 
-        objectToSendToBrowserAction = { message: { modifiers: info.modifiers, word: info.selectionText.trim() }, type: "SearchFromContextMenu", searchEngine: settings["default"] };
+        objectToBeSentToPopupWhenLoaded = { word: info.selectionText.trim(), type: "search_from_context_menu", searchEngine: settings["default"] };
 
         browser.browserAction.openPopup();
-
-    } else if (info.menuItemId == "cm_onpage") {
-        sendWordToTab(info, tab);
-    }
+    } else if (info.menuItemId == "cm_onpage")
+        sendContextMenuSelectedTextToTab(info, tab);
 }
 
-let parentContextMenuOptions = {
-    id: "ContextMenuParent",
-    title: "Seçili kelimeyi çevir",
-    contexts: ["selection"],
-    // throws error in chrome. "Unexpected property: 'icons'."
-    icons: { "32": browser.runtime.getURL("icons/32/icon.png") }
-};
+function createContextMenuEntries() {
+    let parentContextMenuOptions = {
+        id: "ContextMenuParent",
+        title: "Seçili kelimeyi çevir",
+        contexts: ["selection"],
+        // throws error in chrome. "Unexpected property: 'icons'."
+        icons: { "32": browser.runtime.getURL("icons/32/icon.png") }
+    };
 
-//Object.getPrototypeOf(error).name === "TypeError" && error.toString().includes("Unexpected property: 'icons'")
+    try {
+        browser.contextMenus.create(parentContextMenuOptions);
 
-try {
-    browser.contextMenus.create(parentContextMenuOptions);
-} catch (hata) {
-    delete parentContextMenuOptions.icons;
+        console.debug("Successfully created context menu !");
+    } catch (hata) {
+        delete parentContextMenuOptions.icons;
+        try {
+            browser.contextMenus.create(parentContextMenuOptions);
 
-    browser.contextMenus.create(parentContextMenuOptions);
-
-    console.log("Successfully created context menu !");
-}
-
-browser.contextMenus.create({
-    id: "cm_onpage",
-    parentId: "ContextMenuParent",
-    title: "Sayfa üzerinde",
-    contexts: ["selection"]
-});
-
-browser.contextMenus.create({
-    id: "cm_onpopup",
-    parentId: "ContextMenuParent",
-    title: "Uzantı üzerinde",
-    contexts: ["selection"]
-});
-
-browser.contextMenus.onClicked.addListener(ContextMenuClicked);
-
-function sendMessageToTab(message, sender) {
-    browser.tabs.sendMessage(sender.tab.id, message);
-}
-
-// function makeSettingsDefault() {
-//     browser.storage.local.set(defaultSettings);
-//     settings = defaultSettings;
-//     return defaultSettings;
-// }
-
-function makeSettingsDefault(...properties) {
-    if (properties.length === 0) {
-        settings = defaultSettings;
-        browser.storage.local.set(defaultSettings);
-        return defaultSettings;
+            console.debug("Successfully created context menu !");
+        } catch (error) {
+            console.error(error);
+            return;
+        }
     }
 
-    for (property of properties) {
-        if (defaultSettings[property] == null)
-            continue;
+    browser.contextMenus.create({
+        id: "cm_onpage",
+        parentId: "ContextMenuParent",
+        title: "Sayfa üzerinde",
+        contexts: ["selection"]
+    });
 
-        settings[property] = defaultSettings[property];
-
-        let propertyObj = {};
-        propertyObj[property] = defaultSettings[property];
-
-        browser.storage.local.set(propertyObj);
+    if (parentContextMenuOptions.hasOwnProperty("icons")) {
+        browser.contextMenus.create({
+            id: "cm_onpopup",
+            parentId: "ContextMenuParent",
+            title: "Uzantı üzerinde",
+            contexts: ["selection"]
+        });
     }
-
-    return settings;
 }
 
 function updateSettings(e) {
-    let _settings = e;
-
-    //browser.storage.local.onChanged sadece değişen key ve value'yu verdiğinden varsayılan üstüne yazıyoruz.
-
-    for (key in _settings) {
-        settings[key] = _settings[key].newValue;
-    }
-
-    _settings = null;
+    for (const key in e)
+        settings[key] = e[key].newValue;
 }
 
-function checkSettings() {
-    if (!settings) {
-        settings = makeSettingsDefault();
-        console.debug("Settings, nesne boş olduğundan varsayılana ayarlandı.");
+async function searchFromTDK(message) {
+    let word = message.word ?? message;
+
+    let response = await fetch("https://sozluk.gov.tr/gts?ara=" + word, {
+        method: 'GET',
+        credentials: 'omit'
+    }).catch(error => console.error(error));
+
+    if (!response)
+        return;
+    else if (!response.ok) {
+        console.error(response.status, response.statusText);
         return;
     }
 
-    if (!(settings.default === "tdk" || settings.default === "google")) {
-        settings = makeSettingsDefault("default");
-        console.debug("'default' değeri boş olduğundan varsayılana ayarlandı.");
-    }
-
-    if ((Array.isArray(settings.modifiers)) ? settings.modifiers.length === 0 : true) {
-        settings = makeSettingsDefault("modifiers");
-        console.debug("'modifiers' dizisi boş olduğundan varsayılana ayarlandı.");
-    }
-
-    if (typeof settings.searchSelectedTextOnPopupShow !== "boolean") {
-        settings = makeSettingsDefault("searchSelectedTextOnPopupShow");
-        console.debug("'searchSelectedTextOnPopupShow' değeri boş olduğundan varsayılana ayarlandı.");
-    }
-}
-
-function SearchFromTDK(message, sender) {
-    var xhr = new XMLHttpRequest();
-    xhr.onreadystatechange = function () {
-        if (this.readyState != XMLHttpRequest.DONE || !(xhr.status == 0 || (xhr.status >= 200 && xhr.status < 400)))
-            return;
-
-        var list = JSON.parse(xhr.responseText);
-
-        //Hata oluştuysa bitir.
-        if (!list) {
-            console.error("list -> undefined");
-            return;
-        }
-        else if (list.error) {
-            console.error("Sonuç bulunamadı");
-            return;
-        }
-
-        list = list[0];
-        //Anlam var ise devam et.
-        if (parseInt(list.anlam_say) < 1) {
-            return;
-        }
-
-        var anlamlarListe = [];
-        var orneklerListe = [];
-
-        list.anlamlarListe.forEach(a => {
-            anlamlarListe.push({
-                anlam: (a.anlam) ? a.anlam : "",
-                fiil: (a.fiil) ? a.fiil : "",
-                isim: (a.ozelliklerListe) ? (a.ozelliklerListe[0]) ? (a.ozelliklerListe[0].tam_adi) ? a.ozelliklerListe[0].tam_adi : "" : "" : ""
-            });
-            if (a.orneklerListe) {
-                a.orneklerListe.forEach(b => {
-                    orneklerListe.push({
-                        ornek: (b.ornek) ? b.ornek : "",
-                        yazar: (b.yazar) ? (b.yazar[0]) ? (b.yazar[0].tam_adi) ? b.yazar[0].tam_adi : "" : "" : ""
-                    });
-                });
-            }
-        });
-
-        list = null;
-
-        sendMessageToTab({
-            type: "response_from_tdk",
-            anlamlarListe: anlamlarListe,
-            orneklerListe: orneklerListe,
-            x: message.x,
-            y: message.y,
-            word: message.message,
-            isContextMenu: message.isContextMenu
-        }, sender);
-
-        //xhr null yapınca null diye hata veriyor.
-        //xhr = null;
-    }
-    xhr.open("GET", "https://sozluk.gov.tr/gts?ara=" + message.message);
-    xhr.send();
-}
-
-//Sadece 2 cümleye ulaşabiliyoruz.
-function SearchFromGoogle(word) {
-    //"https://www.google.com/search?q=" + word + "+ne+demek"
+    let list;
 
     try {
-        return fetch("https://www.google.com/search?q=" + word + "+ne+demek", {
-            method: 'GET',
-            credentials: 'omit'
-        })
-            .then((response) => response.text())
-            .then((text) => {
-                //Klonlanamaz hatası verdiğinden kapattım
-                //const _document = new DOMParser().parseFromString(text, 'text/html');
-                //return _document;
-
-                return text;
-            }).catch(hata => {
-                console.error(hata);
-            });
+        list = await response.json();
     } catch (error) {
         console.error(error);
     }
+
+    if (!list) {
+        console.error("list -> undefined");
+        return;
+    }
+    else if (list.error) {
+        console.error(list.error);
+        return;
+    }
+
+    list = list[0];
+
+    if (parseInt(list.anlam_say) < 1) {
+        console.error("Anlam listesi boştu.");
+        return;
+    }
+
+    let anlamlarListe = [];
+    let orneklerListe = [];
+
+    list.anlamlarListe.forEach(a => {
+        anlamlarListe.push({
+            anlam: (a.anlam) ? a.anlam : "",
+            fiil: (a.fiil) ? a.fiil : "",
+            isim: (a.ozelliklerListe) ? (a.ozelliklerListe[0]) ? (a.ozelliklerListe[0].tam_adi) ? a.ozelliklerListe[0].tam_adi : "" : "" : ""
+        });
+        if (a.orneklerListe) {
+            a.orneklerListe.forEach(b => {
+                orneklerListe.push({
+                    ornek: (b.ornek) ? b.ornek : "",
+                    yazar: (b.yazar) ? (b.yazar[0]) ? (b.yazar[0].tam_adi) ? b.yazar[0].tam_adi : "" : "" : ""
+                });
+            });
+        }
+    });
+
+    list = null;
+
+    return {
+        type: "response_from_tdk",
+        anlamlarListe: anlamlarListe,
+        orneklerListe: orneklerListe,
+        word: word,
+    };
+}
+
+//Sadece 2 cumleye ulasabiliyoruz.
+async function searchFromGoogle(word) {
+    //"https://www.google.com/search?q=" + word + "+ne+demek"
+
+    let response = await fetch("https://www.google.com/search?q=" + word + "+ne+demek", {
+        method: 'GET',
+        credentials: 'omit'
+    }).catch(err => { console.error(err) });
+
+    if (!response)
+        return;
+    else if (!response.ok) {
+        console.error("Word: " + word, "Result: " + response.status + " " + response.statusText);
+        return;
+    }
+
+    const documentText = await response.text();
+
+    if (!documentText)
+        return;
+
+    const doc = new DOMParser().parseFromString(documentText, 'text/html');
+
+    let query = doc.querySelector("[data-dobid='hdw']");
+
+    if (!query) {
+        console.error("[data-dobid='hdw'] bulunamadı.");
+        return;
+    }
+
+    const wordOnTheDocument = query.textContent;
+
+    if (wordOnTheDocument.trim().toLocaleLowerCase() === "ne demek?") {
+        console.error("Geçersiz kelime seçildiğinden sadece 'ne demek?' cümlesi aratıldı ve elementler eklenmedi.");
+        return;
+    }
+
+    const elements = {};
+
+    const _extraInfoAfterWord = doc.querySelector("ol[class='eQJLDd']").previousElementSibling?.textContent ?? "";
+    const extraInfoAfterWord = (_extraInfoAfterWord) ? ' •' + _extraInfoAfterWord : "";
+
+    let audio = doc.querySelector("audio[jsname='QInZvb']");
+
+    let meaningCounter = 0;
+
+    const contentWrapper = document.createElement('div');
+
+    doc.querySelector("ol[class='eQJLDd']").childNodes.forEach(p => {
+        try {
+            let meaningElement = p.querySelector("div[data-dobid='dfn']");
+
+            if (!meaningElement)
+                return;
+
+            meaningCounter++;
+
+            let extraInfoBeforeExample = meaningElement.parentElement?.previousElementSibling?.textContent ?? "";
+            let extraInfoBeforeExampleElements = (extraInfoBeforeExample) ? [
+                createElement("b", 1, { textContent: extraInfoBeforeExample, style: "font-size: 0.75em" }),
+                createElement("br", 1)
+            ] : null;
+
+            let _similarWord;
+
+            meaningElement.parentElement?.querySelectorAll('div[class^="vmod"]')?.forEach(p => {
+                _similarWord = p.querySelector('div[class*="vmod"]')?.querySelector('div')?.textContent ?? "";
+
+                if (_similarWord && _similarWord.toLocaleLowerCase().includes('benzer:'))
+                    return;
+            });
+
+            let similarWord = (_similarWord && _similarWord.toLocaleLowerCase().includes('benzer:')) ? _similarWord.substring(_similarWord.toLocaleLowerCase().indexOf("benzer:")) : "";
+
+            let similarWordElements;
+
+            if (similarWord)
+                similarWordElements = [
+                    createElement('br', 1),
+                    createElement('b', 1, { textContent: similarWord, className: "ts_similarWord" })
+                ];
+
+            addElements(contentWrapper,
+                extraInfoBeforeExampleElements,
+                createElement("span", 1,
+                    {
+                        textContent: meaningCounter.toString() + ")" + meaningElement.textContent
+                    }),
+                similarWordElements,
+                createElement("br", 2)
+            );
+
+            let exampleElement = meaningElement.nextElementSibling?.className === "vmod" ? meaningElement.nextElementSibling : null;
+
+            if (!exampleElement)
+                return;
+            else if (exampleElement.textContent.trim() === "")
+                return;
+
+            let boldElement = createElement("span", 1, { attributes: { style: "color:rgb(3, 138, 255)" }, "textContent": "Örnek: " });
+            let exampleSpan = createElement("b", 1, { "textContent": exampleElement.textContent });
+
+            console.debug(exampleSpan.textContent);
+
+            addElements(contentWrapper, boldElement, exampleSpan, createElement("br", 2));
+        } catch (err) {
+            console.error(err);
+        }
+    });
+
+    elements.word = word;
+    elements.extraInfoAfterWord = extraInfoAfterWord;
+    elements.audio = audio ? JSON.stringify(audio.innerHTML) : null;
+    elements.contentWrapper = JSON.stringify(contentWrapper.innerHTML);
+
+    return { type: "response_from_google", elements: elements };
+}
+
+function getInformationIfTheWordHasBeenSearchedBefore(word) {
+    if (!word)
+        return;
+
+    let index = 0;
+    for (const obj of settings.storedWords) {
+        if (obj.word === word.trim().toLocaleLowerCase())
+            return { index: index, obj: obj };
+        index++;
+    }
+}
+
+async function search(message) {
+    if (message == null)
+        return;
+
+    const searchEngine = message.searchEngine ?? settings["default"];
+
+    const infoObj = getInformationIfTheWordHasBeenSearchedBefore(message.word);
+    const isItSearchedBefore = infoObj ?? false;
+    const indexOfInfo = (isItSearchedBefore) ? infoObj.index : null;
+    const storedInformation = (isItSearchedBefore) ? infoObj.obj : {};
+
+    if (isItSearchedBefore && storedInformation.hasOwnProperty(searchEngine)) {
+        const info = storedInformation[searchEngine];
+
+        if (info != null)
+            return info;
+    }
+
+    const info = (searchEngine === searchEngineList.tdk)
+        ? await searchFromTDK(message)
+        : await searchFromGoogle(message.word);
+
+    if (!info)
+        return;
+
+    if (isItSearchedBefore) {
+        settings.storedWords[indexOfInfo][searchEngine] = info;
+    } else {
+        settings.storedWords.push({
+            word: message.word,
+            [searchEngine]: info
+        });
+    }
+
+    await browser.storage.local.set({ "storedWords": settings.storedWords });
+
+    return info;
 }
 
 async function onMessage(message, sender) {
-    /*
-    tdk -> message -> {
-        message: "kazı",
-        x: 15,
-        y: 50,
-        type: "search_from_tdk"
-    }
-
-    google -> message -> {
-        message: "kazı",
-        x: 15,
-        y: 50,
-        type: "response_from_google",
-        doc: string HTMLDocument
-    }
-    */
-
-    if (message.type === "search_from_tdk") {
-        SearchFromTDK(message, sender);
-    } else if (message.type === "search_from_google") {
-        let text = await SearchFromGoogle(message.message);
-
-        sendMessageToTab({ type: "response_from_google", doc: text, x: message.x, y: message.y }, sender);
-    }
-
-    return Promise.resolve();
+    if (message.type === "search")
+        return Promise.resolve(await search(message, sender));
 }
 
 function onConnect(port) {
     //browserAction
     if (port.sender.envType == "addon_child") {
-        if (portbA) {
-            portbA.disconnect();
-            portbA = null;
+        if (popupPort) {
+            popupPort.disconnect();
+            popupPort = null;
         }
 
-        //console.debug(port, "browserAction portu bağlandı.");
+        popupPort = port;
 
-        portbA = port;
+        popupPort.name = "bA";
 
-        portbA.name = "bA";
+        if (isWaitingForThePopupToLoadToSendAWord && objectToBeSentToPopupWhenLoaded) {
+            popupPort.postMessage(objectToBeSentToPopupWhenLoaded);
 
-        if (waitingBrowserActionToSendAWord && objectToSendToBrowserAction) {
-            portbA.postMessage(objectToSendToBrowserAction);
-
-            //console.debug(objectToSendToBrowserAction, "mesajı browserAction'a gönderildi.");
-
-            waitingBrowserActionToSendAWord = false;
-            objectToSendToBrowserAction = null;
+            isWaitingForThePopupToLoadToSendAWord = false;
+            objectToBeSentToPopupWhenLoaded = null;
+        }
+        else {
+            popupPort.postMessage({
+                type: "popup_connected_to_background"
+            });
         }
 
-        portbA.onDisconnect.addListener(onDisconnect);
+        popupPort.onDisconnect.addListener(onDisconnect);
 
         return;
     }
 }
 
 function onDisconnect(port) {
-    //console.log(port, "browserAction portunda bağlantı koptu.", "Hata : " + ((port.error) ? port.error : "Yok"));
-
-    console.log(port, "'" + port.name + "' portuyla bağlantı koptu.", (port.error) ? "Error: " + port.error : "");
-
-    if (port.name === "bA") {
-        portbA = null;
-        return;
-    }
+    if (port.name === "bA")
+        popupPort = null;
 }
 
 browser.runtime.onConnect.addListener(onConnect);
@@ -315,3 +351,10 @@ browser.runtime.onConnect.addListener(onConnect);
 browser.runtime.onMessage.addListener(onMessage);
 
 browser.storage.onChanged.addListener(updateSettings);
+
+browser.contextMenus.onClicked.addListener(contextMenuClicked);
+
+createContextMenuEntries();
+
+initializeSettings()
+    .then(_settings => settings = _settings);
