@@ -1,52 +1,12 @@
-/* YAPILACAKLAR
- * 
- */
+import { getSettings, getDefaultSettings } from 'modules/utils.js';
 
-const defaultSettings = Object.freeze({ modifiers: Object.freeze(["none"]), default: "tdk", searchSelectedTextOnPopupShow: false });
-
+let settings = getDefaultSettings();
 let modifierSelectIDCounter = 1;
-
-let blockDiv = document.getElementById("blockDiv");
-
-let selectDiv = document.getElementsByClassName("selectDiv")[0];
-
-let modifierSelects = Object.values(selectDiv.getElementsByTagName("select"));
-
-for (select of modifierSelects) {
-    select.oninput = modifierSelectOnInput;
-}
-
-//img
-let addModifierButton = document.getElementById("addModifierButton");
-
-let searchEngineSelect = document.getElementById("searchEngineSelect");
-searchEngineSelect.oninput = searchEngineSelectOnInput;
-
-function searchSelectedTextOnPopupShowInputOnInput(e) {
-    Save();
-}
-
-let searchSelectedTextOnPopupShowInput = document.getElementById("searchSelectedTextOnPopupShowInput");
-searchSelectedTextOnPopupShowInput.oninput = searchSelectedTextOnPopupShowInputOnInput;
+let blockDiv, selectDiv, modifierSelects, addModifierButton, searchEngineSelect, searchSelectedTextOnPopupShowInput;
+let currentModifierCount = 1;
 
 const modifierCount = 3;
-let currentModifierCount = 1;
 const modifierKeys = Object.freeze(["ctrl", "alt", "shift", "none"]);
-let modifierList = [];
-
-async function loadSettings() {
-    let _settings = await browser.storage.local.get();
-
-    _settings = fixSettings(_settings);
-
-    loadModifierSelectElementsFromSettings(_settings.modifiers);
-
-    searchEngineSelect.value = _settings.default;
-
-    searchSelectedTextOnPopupShowInput.checked = _settings.searchSelectedTextOnPopupShow ?? false;
-}
-
-loadSettings();
 
 function loadModifierSelectElementsFromSettings(modifiers) {
     deleteAllModifierSelectElements();
@@ -55,7 +15,7 @@ function loadModifierSelectElementsFromSettings(modifiers) {
         addModifierSelectElement("none");
     }
 
-    for (modifier of modifiers) {
+    for (let modifier of modifiers) {
         addModifierSelectElement(modifier, true, false);
     }
 }
@@ -63,7 +23,7 @@ function loadModifierSelectElementsFromSettings(modifiers) {
 function deleteAllModifierSelectElements(addDefaultAfterDeletion = false, saveDefaultValue = true) {
     let elements = selectDiv.getElementsByTagName("select");
 
-    for (element of elements) {
+    for (let element of elements) {
         element.remove();
     }
 
@@ -77,63 +37,17 @@ function deleteAllModifierSelectElements(addDefaultAfterDeletion = false, saveDe
     }
 }
 
-function fixSettings(_settings) {
-    let settings = _settings;
-
-    if (!settings) {
-        settings = {};
-        copyObj(defaultSettings, settings);
-        console.debug("Settings, nesne boş olduğundan varsayılana ayarlandı.");
-        return settings;
-    }
-
-    if (!(settings.default === "tdk" || settings.default === "google")) {
-        settings = makeSettingsDefault(settings, "default");
-        console.debug("'default' değeri boş olduğundan varsayılana ayarlandı.");
-    }
-
-    if ((Array.isArray(settings.modifiers)) ? settings.modifiers.length === 0 : true) {
-        settings = makeSettingsDefault(settings, "modifiers");
-        console.debug("'modifiers' dizisi boş olduğundan varsayılana ayarlandı.");
-    }
-
-    if (!(settings.searchSelectedTextOnPopupShow === false || settings.searchSelectedTextOnPopupShow === true)) {
-        settings = makeSettingsDefault(settings, "searchSelectedTextOnPopupShow");
-        console.debug("'searchSelectedTextOnPopupShow' değeri boş olduğundan varsayılana ayarlandı.");
-    }
-
-    return settings;
-}
-
-function makeSettingsDefault(_settings, ...properties) {
-    let settings = _settings;
-
-    if (properties.length === 0) {
-        settings = {};
-        copyObj(defaultSettings, settings);
-    }
-
-    for (property of properties) {
-        if (defaultSettings[property] == null)
-            continue;
-        settings[property] = defaultSettings[property];
-    }
-
-    return settings;
-}
-
 function addModifierSelectElement(modifier, appendChild = true, save = true) {
     if (currentModifierCount === modifierCount) {
         console.debug("Modifier count reached the limit.Which is " + modifierCount);
         return;
     }
-    else if (currentModifierCount > modifierCount)
-    {
+    else if (currentModifierCount > modifierCount) {
         modifierSelects = Object.values(selectDiv.getElementsByTagName("select"));
 
         currentModifierCount = modifierSelects.length;
 
-        for (let i = modifierCount;i < currentModifierCount;i++) {
+        for (let i = modifierCount; i < currentModifierCount; i++) {
             modifierSelects.pop();
         }
     }
@@ -165,11 +79,9 @@ function addModifierSelectElement(modifier, appendChild = true, save = true) {
 }
 
 function addModifierButtonHandler(e) {
-    let _modifiers = [];
+    let _modifiers = [...modifierKeys];
 
-    copyObj(modifierKeys, _modifiers);
-
-    for (el of modifierSelects) {
+    for (let el of modifierSelects) {
         let index = _modifiers.indexOf(el.value);
 
         if (index !== -1)
@@ -189,7 +101,7 @@ function addModifierButtonHandler(e) {
 }
 
 function addModifierOptions(select, modifiers) {
-    for (modifier of modifiers) {
+    for (let modifier of modifiers) {
         if (!modifier)
             continue;
 
@@ -199,23 +111,6 @@ function addModifierOptions(select, modifiers) {
 
         select.appendChild(option);
     }
-}
-
-function copyObj(from, to) {
-    for (let key in from) {
-        let obj = from[key];
-        //We don't want to pass old reference to the array.
-        to[key] = Array.isArray(obj) ? Array.from(obj) : obj;
-    }
-    return to;
-}
-
-function findKeyOfValueInDictionary(dict, value) {
-    for (let key in dict) {
-        if (dict[key] === value)
-            return key;
-    }
-    return null;
 }
 
 function modifierSelectOnInput(e) {
@@ -228,19 +123,19 @@ function modifierSelectOnInput(e) {
     Save();
 }
 
-function searchEngineSelectOnInput(e) {
+function searchEngineSelect_OnInput(e) {
     Save();
 }
 
-function Save() {
+async function Save() {
     console.debug("Save çağırıldı !");
 
     //Prevent user to select anything until the process is done.
     blockDiv.style.visibility = "visible";
 
-    let objects = getSaveObjects();
+    let objects = await getSaveObjects();
 
-    browser.storage.local.set(objects);
+    await browser.storage.local.set(objects);
 
     console.debug("Kaydedilen ayarlar", objects);
 
@@ -251,31 +146,62 @@ function onlyUnique(value, index, self) {
     return self.indexOf(value) === index;
 }
 
-function getSaveObjects() {
-    let objects = {};
-
-    copyObj(defaultSettings, objects);
-
+async function getSaveObjects() {
+    let objects = await browser.storage.local.get();
+    
     modifierSelects = Object.values(selectDiv.getElementsByTagName("select"));
 
-    for (modifierSelect of modifierSelects)
+    objects.modifiers = [];
+
+    for (let modifierSelect of modifierSelects)
         if (modifierSelect.value)
             objects.modifiers.push(modifierSelect.value);
-        else
-            console.debug("Ayarlar kaydedilirken modifierSelect değişkenlerinden birinin \"value\" özelliği boştu.");
 
     objects.default = searchEngineSelect.value;
 
     objects.modifiers = objects.modifiers.filter(onlyUnique);
 
-    objects.searchSelectedTextOnPopupShow = searchSelectedTextOnPopupShowInput.checked;
-
-    //Modifier atanmış işe "none" değeri varsa sil.
-    if (objects.modifiers.length > 1) {
+    if (objects.modifiers.length > 1)
         objects.modifiers = objects.modifiers.filter(modifier => modifier != "none");
-    }
+
+    objects.searchSelectedTextOnPopupShow = searchSelectedTextOnPopupShowInput.checked;
 
     return objects;
 }
 
-addModifierButton.addEventListener("click", addModifierButtonHandler);
+function searchSelectedTextOnPopupShowInput_OnInput(e) {
+    Save();
+}
+
+async function initialize() {
+    settings = await getSettings();
+
+    blockDiv = document.getElementById("blockDiv");
+
+    selectDiv = document.getElementsByClassName("selectDiv")[0];
+
+    modifierSelects = Object.values(selectDiv.getElementsByTagName("select"));
+
+    for (let select of modifierSelects) {
+        select.oninput = modifierSelectOnInput;
+    }
+
+    //img
+    addModifierButton = document.getElementById("addModifierButton");
+
+    searchEngineSelect = document.getElementById("searchEngineSelect");
+    searchEngineSelect.oninput = searchEngineSelect_OnInput;
+
+    searchSelectedTextOnPopupShowInput = document.getElementById("searchSelectedTextOnPopupShowInput");
+    searchSelectedTextOnPopupShowInput.oninput = searchSelectedTextOnPopupShowInput_OnInput;
+
+    loadModifierSelectElementsFromSettings(settings.modifiers);
+
+    searchEngineSelect.value = settings.default;
+
+    searchSelectedTextOnPopupShowInput.checked = settings.searchSelectedTextOnPopupShow ?? false;
+
+    addModifierButton.addEventListener("click", addModifierButtonHandler);
+}
+
+initialize();
